+++ frontend/app/src/entities/proposed-changes/constant.ts
