+++ frontend/app/src/entities/proposed-changes/domain/get-proposed-changes-counts.query.ts
@@ -1,7 +1,7 @@
 import { useCurrentBranch } from "@/entities/branches/ui/branches-provider";
 import { ProposedChangesCountsFromApiParams } from "@/entities/proposed-changes/api/get-proposed-changes-counts-from-api";
+import { PROPOSED_CHANGE_OBJECT } from "@/entities/proposed-changes/constant";
 import { getProposedChangesCounts } from "@/entities/proposed-changes/domain/get-proposed-changes-counts";
-import { PROPOSED_CHANGE_OBJECT } from "@/entities/proposed-changes/utils/constant";
 import { ContextParams, PaginationParams } from "@/shared/api/types";
 import { datetimeAtom } from "@/shared/stores/time.atom";
 import { queryOptions, useQuery } from "@tanstack/react-query";
