+++ frontend/app/src/entities/proposed-changes/utils/computeFilters.ts
@@ -1,13 +1,18 @@
-import { PROPOSED_CHANGE_STATES } from "./constant";
+import { Filter } from "@/shared/hooks/useFilters";
+import { PROPOSED_CHANGE_STATES } from "../constant";
 
-export const computeFilters = ({ filters, state }) => {
+export const computeFilters = ({
+  filters,
+  state,
+}: { filters: Array<Filter>; state: keyof typeof PROPOSED_CHANGE_STATES | string }) => {
   return [
     ...filters,
     {
       name: "state__values",
-      value: state
-        ? (PROPOSED_CHANGE_STATES[state] ?? PROPOSED_CHANGE_STATES.opened)
-        : PROPOSED_CHANGE_STATES.opened,
+      value:
+        state && state in PROPOSED_CHANGE_STATES
+          ? PROPOSED_CHANGE_STATES[state as keyof typeof PROPOSED_CHANGE_STATES]
+          : PROPOSED_CHANGE_STATES.opened,
     },
   ];
 };
